@@ -20,10 +20,13 @@
         <maven.javadoc.failOnError>false</maven.javadoc.failOnError>
         <maven-compiler-plugin.version>3.8.1</maven-compiler-plugin.version>
         <maven-source-plugin.version>3.2.1</maven-source-plugin.version>
-        <spring-boot.version>1.5.9.RELEASE</spring-boot.version>
+        <spring-boot.version>2.6.6</spring-boot.version>
+        <spring-cloud.version>2021.0.1</spring-cloud.version>
         <redisson.version>3.11.0</redisson.version>
         <servlet-api.version>3.0.1</servlet-api.version>
         <ua-util.version>1.21</ua-util.version>
+        <spring.checkstyle.plugin>0.0.23</spring.checkstyle.plugin>
+        <git.commit.plugin>2.2.5</git.commit.plugin>
         <commons-fileupload.version>1.4</commons-fileupload.version>
     </properties>
 
