@@ -5,6 +5,9 @@ package cn.keking.anti_reptile.constant;
  * @since 2019/7/16 10:14
  */
 public class AntiReptileConsts {
-    public static final String VALIDATE_REQUEST_URI = "/anti-reptile/validate";
-    public static final String REFRESH_REQUEST_URI = "/anti-reptile/refresh";
+
+	public static final String VALIDATE_REQUEST_URI = "/anti-reptile/validate";
+
+	public static final String REFRESH_REQUEST_URI = "/anti-reptile/refresh";
+
 }
