@@ -31,91 +31,95 @@ import java.util.regex.Pattern;
  */
 public class AntiReptileInterceptor extends HandlerInterceptorAdapter {
 
+	private String antiReptileForm;
 
-    private String antiReptileForm;
+	private RuleActuator actuator;
 
-    private RuleActuator actuator;
+	private List<String> includeUrls;
 
-    private List<String> includeUrls;
+	private boolean globalFilterMode;
 
-    private boolean globalFilterMode;
+	private VerifyImageUtil verifyImageUtil;
 
-    private VerifyImageUtil verifyImageUtil;
+	private AtomicBoolean initialized = new AtomicBoolean(false);
 
-    private AtomicBoolean initialized = new AtomicBoolean(false);
+	public void init(ServletContext context) {
+		ClassPathResource classPathResource = new ClassPathResource("verify/index.html");
+		try {
+			classPathResource.getInputStream();
+			byte[] bytes = FileCopyUtils.copyToByteArray(classPathResource.getInputStream());
+			this.antiReptileForm = new String(bytes, StandardCharsets.UTF_8);
+		}
+		catch (IOException e) {
+			System.out.println("反爬虫验证模板加载失败！");
+			e.printStackTrace();
+		}
+		ApplicationContext ctx = WebApplicationContextUtils.getWebApplicationContext(context);
+		assert ctx != null;
+		this.actuator = ctx.getBean(RuleActuator.class);
+		this.verifyImageUtil = ctx.getBean(VerifyImageUtil.class);
+		this.includeUrls = ctx.getBean(AntiReptileProperties.class).getIncludeUrls();
+		this.globalFilterMode = ctx.getBean(AntiReptileProperties.class).isGlobalFilterMode();
+		if (this.includeUrls == null) {
+			this.includeUrls = new ArrayList<>();
+		}
+	}
 
-    public void init(ServletContext context) {
-        ClassPathResource classPathResource = new ClassPathResource("verify/index.html");
-        try {
-            classPathResource.getInputStream();
-            byte[] bytes = FileCopyUtils.copyToByteArray(classPathResource.getInputStream());
-            this.antiReptileForm = new String(bytes, StandardCharsets.UTF_8);
-        } catch (IOException e) {
-            System.out.println("反爬虫验证模板加载失败！");
-            e.printStackTrace();
-        }
-        ApplicationContext ctx = WebApplicationContextUtils.getWebApplicationContext(context);
-        assert ctx != null;
-        this.actuator = ctx.getBean(RuleActuator.class);
-        this.verifyImageUtil = ctx.getBean(VerifyImageUtil.class);
-        this.includeUrls = ctx.getBean(AntiReptileProperties.class).getIncludeUrls();
-        this.globalFilterMode = ctx.getBean(AntiReptileProperties.class).isGlobalFilterMode();
-        if (this.includeUrls == null) {
-            this.includeUrls = new ArrayList<>();
-        }
-    }
+	/**
+	 * 拦截器请求前调用
+	 */
+	@Override
+	public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler)
+			throws Exception {
+		if (!initialized.get()) {
+			init(request.getServletContext());
+			initialized.set(true);
+		}
+		HandlerMethod handlerMethod;
+		try {
+			handlerMethod = (HandlerMethod) handler;
+		}
+		catch (ClassCastException e) {
+			return true;
+		}
+		Method method = handlerMethod.getMethod();
+		AntiReptile antiReptile = AnnotationUtils.findAnnotation(method, AntiReptile.class);
+		boolean isAntiReptileAnnotation = antiReptile != null;
+		String requestUrl = request.getRequestURI();
+		if (isIntercept(requestUrl, isAntiReptileAnnotation) && !actuator.isAllowed(request, response)) {
+			CrosUtil.setCrosHeader(response);
+			response.setContentType("text/html;charset=utf-8");
+			response.setStatus(509);
+			VerifyImageDTO verifyImage = verifyImageUtil.generateVerifyImg();
+			verifyImageUtil.saveVerifyCodeToRedis(verifyImage);
+			String str1 = this.antiReptileForm.replace("verifyId_value", verifyImage.getVerifyId());
+			String str2 = str1.replaceAll("verifyImg_value", verifyImage.getVerifyImgStr());
+			String str3 = str2.replaceAll("realRequestUri_value", requestUrl);
+			response.getWriter().write(str3);
+			response.getWriter().close();
+			return false;
+		}
+		return true;
+	}
 
-    /**
-     * 拦截器请求前调用
-     */
-    @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        if (!initialized.get()) {
-            init(request.getServletContext());
-            initialized.set(true);
-        }
-        HandlerMethod handlerMethod;
-        try {
-            handlerMethod = (HandlerMethod) handler;
-        } catch (ClassCastException e) {
-            return true;
-        }
-        Method method = handlerMethod.getMethod();
-        AntiReptile antiReptile = AnnotationUtils.findAnnotation(method, AntiReptile.class);
-        boolean isAntiReptileAnnotation = antiReptile != null;
-        String requestUrl = request.getRequestURI();
-        if (isIntercept(requestUrl, isAntiReptileAnnotation) && !actuator.isAllowed(request, response)) {
-            CrosUtil.setCrosHeader(response);
-            response.setContentType("text/html;charset=utf-8");
-            response.setStatus(509);
-            VerifyImageDTO verifyImage = verifyImageUtil.generateVerifyImg();
-            verifyImageUtil.saveVerifyCodeToRedis(verifyImage);
-            String str1 = this.antiReptileForm.replace("verifyId_value", verifyImage.getVerifyId());
-            String str2 = str1.replaceAll("verifyImg_value", verifyImage.getVerifyImgStr());
-            String str3 = str2.replaceAll("realRequestUri_value", requestUrl);
-            response.getWriter().write(str3);
-            response.getWriter().close();
-            return false;
-        }
-        return true;
-    }
+	/**
+	 * 是否拦截
+	 * @param requestUrl 请求uri
+	 * @param isAntiReptileAnnotation 是否有AntiReptile注解
+	 * @return 是否拦截
+	 */
+	public boolean isIntercept(String requestUrl, Boolean isAntiReptileAnnotation) {
+		if (this.globalFilterMode || isAntiReptileAnnotation || this.includeUrls.contains(requestUrl)) {
+			return true;
+		}
+		else {
+			for (String includeUrl : includeUrls) {
+				if (Pattern.matches(includeUrl, requestUrl)) {
+					return true;
+				}
+			}
+			return false;
+		}
+	}
 
-    /**
-     * 是否拦截
-     * @param requestUrl 请求uri
-     * @param isAntiReptileAnnotation 是否有AntiReptile注解
-     * @return 是否拦截
-     */
-    public boolean isIntercept(String requestUrl, Boolean isAntiReptileAnnotation) {
-        if (this.globalFilterMode || isAntiReptileAnnotation || this.includeUrls.contains(requestUrl)) {
-            return true;
-        } else {
-            for (String includeUrl : includeUrls) {
-                if (Pattern.matches(includeUrl, requestUrl)) {
-                    return true;
-                }
-            }
-            return false;
-        }
-    }
 }
