@@ -8,42 +8,42 @@ import java.io.Serializable;
  */
 public class VerifyImageVO implements Serializable {
 
-    private static final long serialVersionUID = 345634706484343777L;
+	private static final long serialVersionUID = 345634706484343777L;
 
-    private String verifyId;
-    private String verifyType;
-    private String verifyImgStr;
+	private String verifyId;
 
-    public String getVerifyId() {
-        return verifyId;
-    }
+	private String verifyType;
 
-    public void setVerifyId(String verifyId) {
-        this.verifyId = verifyId;
-    }
+	private String verifyImgStr;
 
-    public String getVerifyType() {
-        return verifyType;
-    }
+	public String getVerifyId() {
+		return verifyId;
+	}
 
-    public void setVerifyType(String verifyType) {
-        this.verifyType = verifyType;
-    }
+	public void setVerifyId(String verifyId) {
+		this.verifyId = verifyId;
+	}
 
-    public String getVerifyImgStr() {
-        return verifyImgStr;
-    }
+	public String getVerifyType() {
+		return verifyType;
+	}
 
-    public void setVerifyImgStr(String verifyImgStr) {
-        this.verifyImgStr = verifyImgStr;
-    }
+	public void setVerifyType(String verifyType) {
+		this.verifyType = verifyType;
+	}
 
-    @Override
-    public String toString() {
-        return "VerifyImageVO{" +
-                "verifyId='" + verifyId + '\'' +
-                ", verifyType='" + verifyType + '\'' +
-                ", verifyImgStr='" + verifyImgStr + '\'' +
-                '}';
-    }
+	public String getVerifyImgStr() {
+		return verifyImgStr;
+	}
+
+	public void setVerifyImgStr(String verifyImgStr) {
+		this.verifyImgStr = verifyImgStr;
+	}
+
+	@Override
+	public String toString() {
+		return "VerifyImageVO{" + "verifyId='" + verifyId + '\'' + ", verifyType='" + verifyType + '\''
+				+ ", verifyImgStr='" + verifyImgStr + '\'' + '}';
+	}
+
 }
