@@ -10,28 +10,29 @@ import java.util.List;
  */
 public class RuleActuator {
 
-    private  List<AntiReptileRule> ruleList;
+	private List<AntiReptileRule> ruleList;
 
-    public RuleActuator(List<AntiReptileRule> rules) {
-        ruleList = rules;
-    }
+	public RuleActuator(List<AntiReptileRule> rules) {
+		ruleList = rules;
+	}
 
-    /**
-     * 是否允许通过请求
-     * @param request 请求
-     * @param response 响应
-     * @return 请求是否允许通过
-     */
-    public boolean isAllowed(HttpServletRequest request , HttpServletResponse response){
-        for (AntiReptileRule rule: ruleList){
-            if (rule.execute(request,response)){
-                return false;
-            }
-        }
-        return true;
-    }
+	/**
+	 * 是否允许通过请求
+	 * @param request 请求
+	 * @param response 响应
+	 * @return 请求是否允许通过
+	 */
+	public boolean isAllowed(HttpServletRequest request, HttpServletResponse response) {
+		for (AntiReptileRule rule : ruleList) {
+			if (rule.execute(request, response)) {
+				return false;
+			}
+		}
+		return true;
+	}
 
-    public void reset(HttpServletRequest request, String realRequestUri){
-       ruleList.forEach(rule -> rule.reset(request, realRequestUri));
-    }
+	public void reset(HttpServletRequest request, String realRequestUri) {
+		ruleList.forEach(rule -> rule.reset(request, realRequestUri));
+	}
+
 }
